LPD-62959 Create new project for fragments tests
