+++ modules/apps/layout/layout-content-page-editor-web/test.properties
@@ -13,6 +13,7 @@ modules.includes.required.test.batch.class.names.includes[modules-unit][relevant
 
 playwright.projects.includes[playwright-js-tomcat101-mysql57][relevant][layout-content-page-editor-web-playwright-rule]=\
     layout-content-page-editor-web.form-container,\
+    layout-content-page-editor-web.fragments,\
     layout-content-page-editor-web.main
 
 relevant.rule.names=\
