+++ modules/test/playwright/tests/layout-content-page-editor-web/fragments/collectionDisplayFragment.spec.ts
@@ -5,22 +5,22 @@
 
 import {expect, mergeTests} from '@playwright/test';
 
-import {apiHelpersTest} from '../../../../fixtures/apiHelpersTest';
-import {collectionsPagesTest} from '../../../../fixtures/collectionsPagesTest';
-import {featureFlagsTest} from '../../../../fixtures/featureFlagsTest';
-import {isolatedSiteTest} from '../../../../fixtures/isolatedSiteTest';
-import {loginTest} from '../../../../fixtures/loginTest';
-import {pageEditorPagesTest} from '../../../../fixtures/pageEditorPagesTest';
-import {pageManagementSiteTest} from '../../../../fixtures/pageManagementSiteTest';
-import {pageViewModePagesTest} from '../../../../fixtures/pageViewModePagesTest';
-import {clickAndExpectToBeHidden} from '../../../../utils/clickAndExpectToBeHidden';
-import {clickAndExpectToBeVisible} from '../../../../utils/clickAndExpectToBeVisible';
-import {getContentSetElementContent} from '../../../../utils/getContentSetElementContent';
-import getRandomString from '../../../../utils/getRandomString';
-import {ANIMALS_COLLECTION_NAME} from '../../../setup/page-management-site/main/constants/animals';
-import getCollectionDefinition from '../utils/getCollectionDefinition';
-import getFragmentDefinition from '../utils/getFragmentDefinition';
-import getPageDefinition from '../utils/getPageDefinition';
+import {apiHelpersTest} from '../../../fixtures/apiHelpersTest';
+import {collectionsPagesTest} from '../../../fixtures/collectionsPagesTest';
+import {featureFlagsTest} from '../../../fixtures/featureFlagsTest';
+import {isolatedSiteTest} from '../../../fixtures/isolatedSiteTest';
+import {loginTest} from '../../../fixtures/loginTest';
+import {pageEditorPagesTest} from '../../../fixtures/pageEditorPagesTest';
+import {pageManagementSiteTest} from '../../../fixtures/pageManagementSiteTest';
+import {pageViewModePagesTest} from '../../../fixtures/pageViewModePagesTest';
+import {clickAndExpectToBeHidden} from '../../../utils/clickAndExpectToBeHidden';
+import {clickAndExpectToBeVisible} from '../../../utils/clickAndExpectToBeVisible';
+import {getContentSetElementContent} from '../../../utils/getContentSetElementContent';
+import getRandomString from '../../../utils/getRandomString';
+import {ANIMALS_COLLECTION_NAME} from '../../setup/page-management-site/main/constants/animals';
+import getCollectionDefinition from '../main/utils/getCollectionDefinition';
+import getFragmentDefinition from '../main/utils/getFragmentDefinition';
+import getPageDefinition from '../main/utils/getPageDefinition';
 
 const test = mergeTests(
 	apiHelpersTest,
