+++ modules/test/playwright/tests/layout-content-page-editor-web/fragments/config.ts
@@ -1,0 +1,10 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+export const config = {
+	dependencies: ['page-management-site.main'],
+	name: 'layout-content-page-editor-web.fragments',
+	testDir: 'tests/layout-content-page-editor-web/fragments',
+};
