+++ modules/test/playwright/tests/layout-content-page-editor-web/fragments/test.properties
@@ -1,0 +1,1 @@
+testray.main.component.name=Content Pages
