+++ test.properties
@@ -4660,6 +4660,7 @@
         iframe-web.main,\
         layout-admin-web.main,\
         layout-content-page-editor-web.form-container,\
+        layout-content-page-editor-web.fragments,\
         layout-content-page-editor-web.main,\
         layout-locked-layouts-web.main,\
         layout-page-template-admin-web.main,\
